requirements.txt
qrcode==7.2
Flask==2.0.1
Pillow==8.3.1
